Fix warning in GG_String::cmpFirstWord(char *, char *)
Avoids the GCC and Clang warning:

gg_string.cpp: In static member function ‘static bool GG_String::cmpFirstWord(char*, char*)’:
gg_string.cpp:202:25: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

Whilst here, improve the documentation for GG_String::cmpFirstWord().

+++ src/gamegui/gg_string.cpp
@@ -191,13 +191,20 @@ bool GG_String::getQuote( char *str, char *quote )
 }
 
 
-//  Skip whitespace in str, compare the following characters with
-//  null-terminated wrd, return true on exact match
-//
-bool GG_String::cmpFirstWord( char *str, char *wrd )
+/**
+ * Skip whitespace in str, compare the following characters with
+ * null-terminated wrd
+ *
+ * @param str String to skip whitespace
+ * @param wrd String to compare against
+ *
+ * @return True on exact match
+ */
+bool
+GG_String::cmpFirstWord( char *str, char *wrd )
 {
   char *p = skipWhiteSpace(str);
-  int  wrdLen = strlen(wrd);
+  size_t  wrdLen = strlen(wrd);
 
   if( wrdLen <= strlen(p) )
     if( memcmp( p, wrd, wrdLen ) == 0 )
